Compute dashboard metrics from requirement data
