+++ src/components/Dashboard.tsx
@@ -1,3 +1,4 @@
+import { useMemo } from "react";
 import type { Requirement } from "../types";
 
 import {
@@ -12,12 +13,23 @@ const COLORS = ["#0097D5", "#00C49F", "#FFBB28", "#FF8042", "#8884D8"];
 
 interface Props {
   requirements: Requirement[];
-  coveragePercent: number;
-  isReady: boolean;
   statuses: string[];
 }
 
-export function Dashboard({ requirements, coveragePercent, isReady, statuses }: Props) {
+export function Dashboard({ requirements, statuses }: Props) {
+  const coveragePercent = useMemo(() => {
+    if (requirements.length === 0) return 0;
+    const verified = requirements.filter(
+      (r) => r.status === "verified" || r.status === "closed"
+    ).length;
+    return Math.round((verified / requirements.length) * 100);
+  }, [requirements]);
+
+  const isReady = useMemo(
+    () => requirements.every((r) => r.status === "verified" || r.status === "closed"),
+    [requirements]
+  );
+
   const pieData = statuses.map((s) => ({
     name: s,
     value: requirements.filter((r) => r.status === s).length,
